depricated/pyproject.toml
[tool.poetry]
package-mode = false
name = "fastapi-backend"
version = "0.0.0"
description = "FastAPI pet project"
authors = ["Willy Wonka <arzhevitin.g@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
SQLAlchemy = "^2.0.31"
alembic = "^1.13.2"
loguru = "^0.7.2"
pyjwt = "^2.8.0"
bcrypt = "4.1.3"
redis = {extras = ["hiredis"], version = "^5.0.7"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
